Collect Token Cleanup
Removed a require function from the collectTokens function that no longer made sense, given we can add tokens to the guild bank assuming. Also added a check that, if the whitelisted token has a zero balance, adding it won't cause the total guild tokens to exceed the max. Also removed the onlyDelegate modifier, so that it could be used for payments.